server/requirements.txt
Django==5.0.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow==10.4.0
django-filter==24.2
python-decouple==3.8
setuptools<81
